refactor: 마감일 계산 함수 유틸로 분리
